Clean up #5104 based on reviews
Add back train_asr.yaml symlink
Fix permissions
Move ugly data configs to https://github.com/tjysdsg/AphasiaBank_config

[pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci
