+++ egs2/aphasiabank/asr1/conf/train_asr.yaml
@@ -1,0 +1,1 @@
+tuning/train_asr_ebranchformer_small_wavlm_large1_interctc6.yaml
