+++ egs2/aphasiabank/asr1/local/score_aphasia_detection.py
@@ -5,7 +5,7 @@ Calculate sentence-level and speaker-level Aphasia detection accuracy
 import argparse
 from collections import Counter
 
-from config import spk2aphasia_label, utt2spk
+from data import spk2aphasia_label, utt2spk
 
 
 def get_args():
