+++ egs2/aphasiabank/asr1/local/score_per_severity.sh
