+++ egs2/aphasiabank/asr1/run.sh
@@ -6,28 +6,23 @@ set -e
 set -u
 set -o pipefail
 
-asr_tag="ebranchformer_wavlm"
-
 train_set="train"
 valid_set="val"
 test_sets="test"
 include_control=true
 tag_insertion=none
 
-asr_config=conf/tuning/train_asr_ebranchformer_small_wavlm_large1.yaml
+asr_config=conf/train_asr.yaml
 inference_config=conf/decode.yaml
 
 feats_normalize=global_mvn
-if [[ ${asr_config} == *"hubert"* ]] || [[ ${asr_config} == *"wavlm"* ]]; then
+config_name=$(readlink -f ${asr_config})
+if [[ ${config_name} == *"hubert"* ]] || [[ ${config_name} == *"wavlm"* ]]; then
   feats_normalize=utt_mvn # https://github.com/espnet/espnet/issues/4006#issuecomment-1047898558
 fi
 
 ./asr.sh \
-  --asr_tag "${asr_tag}" \
   --lang en \
-  --inference_nj 2 \
-  --gpu_inference true \
-  --ngpu 2 \
   --max_wav_duration 33 \
   --audio_format wav \
   --feats_type raw \
