+++ egs2/aphasiabank/asr1/run_interctc.sh
@@ -9,7 +9,6 @@ set -e
 set -u
 set -o pipefail
 
-asr_tag="ebranchformer_wavlm_interctc6"
 train_set="train"
 valid_set="val"
 test_sets="test"
@@ -20,16 +19,13 @@ asr_config=conf/tuning/train_asr_ebranchformer_small_wavlm_large1_interctc6.yaml
 inference_config=conf/decode.yaml
 
 feats_normalize=global_mvn
-if [[ ${asr_config} == *"hubert"* ]] || [[ ${asr_config} == *"wavlm"* ]]; then
+config_name=$(readlink -f ${asr_config})
+if [[ ${config_name} == *"hubert"* ]] || [[ ${config_name} == *"wavlm"* ]]; then
   feats_normalize=utt_mvn # https://github.com/espnet/espnet/issues/4006#issuecomment-1047898558
 fi
 
 ./asr.sh \
-  --asr_tag "${asr_tag}" \
   --lang en \
-  --inference_nj 2 \
-  --gpu_inference true \
-  --ngpu 2 \
   --max_wav_duration 33 \
   --audio_format wav \
   --feats_type raw \
